Ship lxtrace like it were firmware
The spectrum kernel module build also produces a userspace tool that is
tightly coupled with the kernel version. It is called lxtrace-$(uname -r)
and it is needed whenever we need to debug gpfs at lower levels.
`mmtracectl --start` is the command to check this.

So we copy this lxtrace-$(uname-r) binary in /opt/lxtrace during the KMM
build, the KMM operator will place it under /var/lib/firmware on the
nodes and then in the buildgpl configmap we take it from
/var/lib/firmware/ and copy it from /host/var/lib/firmware/lxtrace-*
to /usr/lpp/mmfs/bin/