Always pull kmm image
Use case here is if a user deploys fusion without secret signing and
then later wants to retrigger a build with the secret for signing
modules in place. Without this, you have to disable the kmm pods, go on
every node and remove the existing kmm build image via crictl commands.