Stop using the imagehash as selector for the kmm module
At the time we introduced this selector in order to add some guarantees
that a new gpfs module would only be loaded on nodes which already have
a new core pod.

The problem with this is that if a pod restarts it will briefly remove
that image-digest label, which will trigger kmm to try and rmmod the
kernel module. This will fail of course, because the filesystem is still
mounted.

Let's just use the daemon-selector which is a label that does not go
away when core pods are restarted.

The risk of getting new gpfs kernel modules on nodes which have old CNSA
is next to zero because kmm will be unable to rmmod the kernel modules
anyways because the core pods are running.

Co-Authored-By: Akos Eros <aeros@redhat.com>